+++ tests/testthat/test-mock.R
@@ -1,4 +1,6 @@
-test_that("deprecated", {
-  expect_snapshot(local_mock())
-  expect_snapshot(with_mock(is_testing = function() FALSE))
+test_that("now defunct", {
+  expect_snapshot(error = TRUE, {
+    local_mock()
+    with_mock(is_testing = function() FALSE)
+  })
 })